[Broker/Market] Improve price handling and position management
[Description]
- Modified price retrieval functions to return None instead of 0 for invalid prices
- Added explicit checks for None prices throughout trading workflow
- Enhanced casting/size handling for order positions
- Improved error logging for broker interactions
- Updated portfolio calculation to skip positions with missing prices
- Modified strategy sizing calculations to use integer positions

[Motivation]
- Prevent invalid trading decisions caused by using 0 as valid price
- Avoid crashes when broker communication fails
- Required to handle real-world trading scenarios with unreliable market data
- Fixes issues with partial shares and non-integer position sizing
- Ensures portfolio valuations exclude assets with unknown prices

[Self-Test]
- Verified broker integration test suite passes
- Manual market data retrieval in different error states
- Tested portfolio valuation with partially missing prices
- Confirmed integer position handling in mock trades
- Tested order handling with non-integer size data parsing

+++ strategy/basic/movingprofit.py
@@ -98,7 +98,7 @@ class MovingProfitStrategy(BasicExitStrategy):
                     lot_size = pos.size / self.LOT_UNIT
                     sell_pos = 0
                     if lot_size >= 2 and floor(lot_size/2) * self.LOT_UNIT * price >= self.MIN_CASH_PER_TRADE:
-                        sell_pos = ceil(lot_size/2) * self.LOT_UNIT
+                        sell_pos = int(ceil(lot_size/2) * self.LOT_UNIT)
                     else:
                         sell_pos = lot_size * self.LOT_UNIT
                     # dbg_info(f'Selling debug: {pos.size}->{sell_pos}')
@@ -117,7 +117,7 @@ class MovingProfitStrategy(BasicExitStrategy):
                     lot_size = pos.size / self.LOT_UNIT
                     sell_pos = 0
                     if lot_size >= 2 and floor(lot_size/2) * self.LOT_UNIT * price >= self.MIN_CASH_PER_TRADE:
-                        sell_pos = ceil(lot_size/2) * self.LOT_UNIT
+                        sell_pos = int(ceil(lot_size/2) * self.LOT_UNIT)
                     else:
                         sell_pos = lot_size * self.LOT_UNIT
                     # dbg_info(f'Selling debug: {pos.size}->{sell_pos}')
